expose dleq_prove and dleq_verify
These functions in the ecdsa_adaptor module are static/internal. We
want to make use of them in the BitBox firmware, so we expose them.

They are useful in the context of silent payments (BIP-352). In the
future, we expect new BIPs for silent payments in PSBT including a
DLEQ specification, and a dedicated silent payment module in
libsecp256k1 that includes the DLEQ functions. Until then, we use the
ones here.

See also:

- https://delvingbitcoin.org/t/bip352-psbt-support/877
- https://gist.github.com/andrewtoth/df97c3260cc8d12f09d3855ee61322ea
- https://github.com/bitcoin-core/secp256k1/pull/1519